Add Repl tests
